Change type of talsz and dependent code from size_t to int. Tal ids are already stored as int and and talsz is the limit for these ids. OK tb@

+++ usr.sbin/rpki-client/extern.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: extern.h,v 1.126 2022/04/12 08:45:34 tb Exp $ */
+/*	$OpenBSD: extern.h,v 1.127 2022/04/19 09:52:29 claudio Exp $ */
 /*
  * Copyright (c) 2019 Kristaps Dzonsons <kristaps@bsd.lv>
  *
@@ -414,7 +414,7 @@ extern int filemode;
 extern const char *tals[];
 extern const char *taldescs[];
 extern unsigned int talrepocnt[];
-extern size_t talsz;
+extern int talsz;
 
 /* Routines for RPKI entities. */
